Refactoring of code structure
