+++ message/validation/alan/common_checks.go
@@ -1,0 +1,60 @@
+package msgvalidation
+
+import (
+	"fmt"
+	"time"
+
+	"github.com/attestantio/go-eth2-client/spec/phase0"
+	spectypes "github.com/bloxapp/ssv-spec/alan/types"
+	"golang.org/x/exp/slices"
+)
+
+// ownCommittee should be called only when WithOwnOperatorID is set
+func (mv *messageValidator) ownCommittee(committee []spectypes.OperatorID) bool {
+	return slices.ContainsFunc(committee, func(operatorID spectypes.OperatorID) bool {
+		return operatorID == mv.operatorDataStore.GetOperatorID()
+	})
+}
+
+func (mv *messageValidator) committeeRole(role spectypes.RunnerRole) bool {
+	return role == spectypes.RoleCommittee
+}
+
+func (mv *messageValidator) validateSlotTime(messageSlot phase0.Slot, role spectypes.RunnerRole, receivedAt time.Time) error {
+	if mv.earlyMessage(messageSlot, receivedAt) {
+		return ErrEarlyMessage
+	}
+
+	if lateness := mv.lateMessage(messageSlot, role, receivedAt); lateness > 0 {
+		e := ErrLateMessage
+		e.got = fmt.Sprintf("late by %v", lateness)
+		return e
+	}
+
+	return nil
+}
+
+func (mv *messageValidator) earlyMessage(slot phase0.Slot, receivedAt time.Time) bool {
+	return mv.netCfg.Beacon.GetSlotEndTime(mv.netCfg.Beacon.EstimatedSlotAtTime(receivedAt.Unix())).
+		Add(-clockErrorTolerance).Before(mv.netCfg.Beacon.GetSlotStartTime(slot))
+}
+
+func (mv *messageValidator) lateMessage(slot phase0.Slot, role spectypes.RunnerRole, receivedAt time.Time) time.Duration {
+	var ttl phase0.Slot
+	switch role {
+	case spectypes.RoleProposer, spectypes.RoleSyncCommitteeContribution:
+		ttl = 1 + lateSlotAllowance
+	case spectypes.RoleCommittee, spectypes.RoleAggregator:
+		ttl = 32 + lateSlotAllowance
+	case spectypes.RoleValidatorRegistration, spectypes.RoleVoluntaryExit:
+		return 0
+	default:
+		panic("unexpected role")
+	}
+
+	deadline := mv.netCfg.Beacon.GetSlotStartTime(slot + ttl).
+		Add(lateMessageMargin).Add(clockErrorTolerance)
+
+	return mv.netCfg.Beacon.GetSlotStartTime(mv.netCfg.Beacon.EstimatedSlotAtTime(receivedAt.Unix())).
+		Sub(deadline)
+}
