+++ message/validation/alan/const.go
@@ -1,0 +1,23 @@
+package msgvalidation
+
+import (
+	"time"
+)
+
+const (
+	// lateMessageMargin is the duration past a message's TTL in which it is still considered valid.
+	lateMessageMargin = time.Second * 3
+
+	// clockErrorTolerance is the maximum amount of clock error we expect to see between nodes.
+	clockErrorTolerance = time.Millisecond * 50
+
+	maxMessageSize             = maxConsensusMsgSize // TODO: calculate new value
+	maxConsensusMsgSize        = 8388608
+	maxPartialSignatureMsgSize = 1952
+	allowedRoundsInFuture      = 1
+	allowedRoundsInPast        = 2
+	lateSlotAllowance          = 2
+	rsaSignatureSize           = 256
+	blsSignatureSize           = 96
+	maxDutiesPerEpoch          = 2
+)
